Basic configurtion system

+++ src/main.rs
@@ -1,12 +1,14 @@
 mod app;
 mod tui;
 mod ui;
+mod json;
 
 use crate::app::App;
+use crate::json::{Colors, read_and_extract_json};
 use crate::tui::Tui;
 
-use std::env::{args, current_dir};
-use std::fs::{File, read_to_string};
+use std::env::{args, current_dir, var};
+use std::fs::{File, read_to_string, exists};
 use std::io::Write;
 
 use crossterm::event::KeyModifiers;
@@ -23,21 +25,48 @@ fn main() -> color_eyre::Result<()> {
     
     let file_name = args[1].clone();
     let file_path = format!("{}/{}", cwd, file_name);
-    let file_content = read_to_string(&file_path)?;
 
-    let mut file  = File::create(&file_path)?;
+    let mut file_content = "".to_string();
+    let mut file: File;
+
+    if exists(&file_path)? {
+        file_content = read_to_string(&file_path)?;
+        file  = File::create(&file_path)?;
+    } else {
+        file  = File::create(&file_path)?;
+    }
+    
     let mut app = App::new(file_name, file_path, &file_content);
     
     let backend = CrosstermBackend::new(std::io::stderr());
     let terminal = Terminal::new(backend)?;
     let mut tui = Tui::new(terminal);
+
+    let config_path_variable = "KRUSTE_CONFIG";
+    match var(config_path_variable) {
+        Ok(value) => {
+            println!("Found Kruste config at {}.", value);
+            let config_file = value;
+            app.colors = read_and_extract_json(config_file);
+        }
+        Err(e) => {
+            println!("Error: Couldn't find Kruste config!\n{}: {}", config_path_variable, e);
+            println!("Using default config for now.");
+            app.colors = Colors {
+                text: "#FFFFFF".to_string(),
+                background: "#000000".to_string(),
+                border: "#FFFFFF".to_string()
+            }
+        }
+    }
+
     tui.enter()?;
 
     while !app.should_quit {
         let current_editor_content = app.textarea.lines().join("\n");
 
         if app.ask_save {
-            tui.draw_ask_save()?;
+            tui.draw_ask_save(&mut app)?;
 
             if let Event::Key(key) = read()? {
                 match key.code {
@@ -58,6 +87,7 @@ fn main() -> color_eyre::Result<()> {
                         if file_content != app.textarea.lines().join("\n") {
                             app.ask_save = true;
                         } else {
+                            file.write_all(format!("{}", file_content).as_bytes())?;
                             app.quit();
                         }
                     }
